+++ src/commands/auth.ts
@@ -1,7 +1,7 @@
 import { commands, env, ExtensionContext, Uri } from "vscode";
 import { EXTENSION_ID } from "../constants";
 import { signIn, signout } from "../store/auth";
-import { GistsNode } from "../tree/nodes";
+import { FollowedUserGistsNode, GistsNode } from "../tree/nodes";
 
 export async function registerAuthCommands(context: ExtensionContext) {
   context.subscriptions.push(
@@ -14,8 +14,10 @@ export async function registerAuthCommands(context: ExtensionContext) {
   context.subscriptions.push(
     commands.registerCommand(
       `${EXTENSION_ID}.openProfile`,
-      async (node: GistsNode) => {
-        const uri = Uri.parse(`https://gist.github.com/${node.login}`);
+      async (node: GistsNode | FollowedUserGistsNode) => {
+        const login =
+          node instanceof GistsNode ? node.login : node.user.username;
+        const uri = Uri.parse(`https://gist.github.com/${login}`);
         env.openExternal(uri);
       }
     )