+++ src/extension.ts
@@ -8,7 +8,7 @@ import { store } from "./store";
 import { initializeAuth } from "./store/auth";
 import { initializeStorage } from "./store/storage";
 import { registerTreeProvider } from "./tree";
-import { getGistWorkspaceId, isGistWorkspace, openGist } from "./utils";
+import { getGistWorkspaceId, isGistWorkspace, openGistFiles } from "./utils";
 
 export async function activate(context: vscode.ExtensionContext) {
   log.setLoggingChannel(vscode.window.createOutputChannel("GistPad"));
@@ -21,7 +21,7 @@ export async function activate(context: vscode.ExtensionContext) {
 
   if (isGistWorkspace()) {
     const gistId = getGistWorkspaceId();
-    openGist(gistId, false);
+    openGistFiles(gistId);
   }
 
   initializeStorage(context);